+++ bika/lims/content/analysisrequest.py
@@ -2546,10 +2546,20 @@ class AnalysisRequest(BaseFolder):
         """Obtains the sampling round UID
         :returns: UID
         """
-        if self.getSamplingRound():
-            return self.getSamplingRound().UID()
+        sr = self.getSamplingRound()
+        if sr:
+            return sr.UID()
         else:
             return ''
+
+    def getStorageLocationTitle(self):
+        """ A method for AR listing catalog metadata
+        :return: Title of Storage Location
+        """
+        sl = self.getStorageLocation()
+        if sl:
+            return sl.Title()
+        return ''
 
     @security.public
     def getResultsRange(self):